Remove NULL type since it doesn't have meaning here
When it's possible to store different types of data in the same row,
the NULL value is important.  However, this doesn't map to anything
in bloop because an attribute will be None to signal missing.

Because a Column of type NULL can only ever store a literal 'true'
and never changes (it can't lack a value, for instance) it is
effectively half of a boolean.