_BaseModel __init__ doesn't set missing columns to None #23
This is the first significant deviation of the overhaul to tracking.
Because tracking changes any time a field is set, there were two
(at least) options.

1. Mark tracking on every __set__ and __del__
2. Exclude certain set/del from marking (such as during __init__)

The first requires that we don't set missing values at all, even to
None.  Doing so would indicate an explicit lack of value.
The second allows us to set Nones for lack of values, using (for
instance) a special context that indicates we aren't tracking these
changes.  This adds complexity to any code that loads/dumps values,
requiring they remember the magic context to disable tracking.

With the first, behavior (and documentation) can be very clear -
don't set a value unless it's not empty, or is expected to be
None when persisted.