Add basic unit tests for tracking #23
There's an odd coverage issue with the continue on hash/range key
columns in tracking diff generation.  I've added an extensive comment
as well as sources that verify the behavior, in case someone stumbles
upon the "skip coverage of this branch entirely" and panics.

Column's __del__ will always mark for deletion, even if the attribute
didn't exist locally.  In such a case it still raises AttributeError;
I should revisit this decision in the morning.  I *think* it should
be enough for callers to set the attribute to None to indicate removal,
without needing to check for existence first (as they'd need to do with
del).