Overhaul Engine.load, streamline tracking interface
First pass at breaking out the phases of a Engine.load call.

Tracking interface has added engine as a required parameter in a
few places, mainly around touching the atomic snapshot.  This makes
it possible to pass different engines (or EngineViews) when
manipulating a snapshot, as well as moving the decision to the
tracking module itself.

Previously, callers would decide if it was appropriate to modify
the snapshot at all.  Now, the tracking module can determine when
it is appropriate to modify or clear the snapshot.