Instead of storing tracking data ON the object, use a WeakKeyDictionary
This simplifies introspection (there isn't a *giant* blob hanging around
when looking at a model instance __dict__) and keeps the data contained
within the `tracking.py` module.

Next up for consideration: because the weakref dict will automatically
clean up data for us when garbage collection occurs, is it reasonable
to re-enable auto atomic condition caching?  This would remove the
restriction on atomic save/delete that requires the object to have
been loaded atomically.  The increase in time/space is still a
concern but the lifetime of the extraneous data is no longer unbounded.