Re-enable atomic caching
This was originally disabled due to concerns around the tracking module's
per-object state management.  With the introduction of a WeakRefDictionary
however, these concerns are greatly reduced.  The overhead is minimal for
normal usage; loading thousands of objects at once may incur a nominal
performance penalty.

If this becomes an issue, I suspect the tracking module can pivot to
take the already-dumped wire format, instead of re-dumping through
the engine again.  This would almost entirely eliminate the overhead.

The biggest win is ease of use; the last refactor required an object
to be loaded through an atomic engine just to use atomic for save/delete.
This required passing the atomic-ness of an object with it wherever it went,
for fear of trying to save against a None atomic condition.