Drop Engine.model, Engine.bind requires BaseModel #32
This is one of two big refactors.  This fully disconnects engines
and models.  The underlying TypeEngine still manages its own
associations through binding typedefs, but it's now possible to
bind multiple base models to a single engine, and multiple engines
to one base model.

This makes the setup process one line longer, and requires a reference
to the BaseModel at engine bind, but removes most of the state that
and engine was required to track.

The docs are now out of date in a *lot* of places.
The ones I've found with 'git grep -n "\.model"' that aren't
in this commit:
* users/advanced
* users/engine
* users/models
* users/types