Drop engine context
EngineView was a useful construct when models were bound to engines,
but much less so now that the types are only bound to the type_engine.
Since Engine's __init__ takes type_engine now, there's no need to
wrap an engine anymore.  The context mechanism will almost certainly
break for any engine subclasses, since it won't carry custom values
forward.

Instead, there's now an example of copying an engine with slightly
different config in the Common Patterns section of the docs.