Part 1 of ConditionRenderer refactor
The goal is to centralize most of the decision making when rendering.
A caller should be able to pass the kwargs to `render` that they want
included, and the ConditionRenderer is invoked correctly.

Tests will change around as the order of rendering gets shuffled.
The conditions are effectively; the attribute names and values are
simply in a different order.