Part 2 of ConditionRenderer refactor
This time refactoring Filter.build to use render instead of
ConditionRenderer.  This required adding a _select_attributes
field to the Filter instance, and slightly massaging the validation.

There are still some rough edges around the renderer, filters, and
conditions.  I need to come up with a way to split the renderer out
of the condition tests (mocking is good!) and take another look at
Filter's interface.  I like the chained style, but request's
Session configuration model is appealing (standard properties).
It would reduce some of the "is this a function or an attribute"
confusion, and in many cases take up as many lines when building
queries.