Map.dynamo_load calls dynamo_load(None) on inner types for None #49
Without this, Map.dynamo_load would return an empty dict, which
is too strong an assumption for the path keys.  With an empty dict,
an inner custom type could map None to a value (say, empty set)
which would be skipped by Map's dynamo_load.

Map is unlike Set, List, and TypedMap, which have an unbound size and
cannot know how many Nones to pass to their inner types.