Type._dump returns raw None instead of {str: None} #49
- drop None guards from tracking.sync
- simplify not_none in all vector type dynamo_dump functions
- dumping None through a model returns None
- new VectorModel in test models

Aside from documentation, tests need to be written with the
new VectorModel to check conditions, atomics, and the update builder
to ensure they handle things correctly.  The answer *should* be yes,
since _dump's only new behavior is sometimes returning None.  That
may simplify logic, but it shouldn't require a refactor.

Now, if there's any lingering code that calls _load, that may get
objects (list, set) instead of None due to the lack of guards; that'd
be a problem, especially since the tests are passing right now.