Add unit tests for BaseModel.__str__, drop BaseModel.__eq__
There are some different meanings for "==" now that a new instance
of a model has missing attributes, while a loaded may have None,
or an empty object like `set()`.  This makes the previous
getattr(obj, name, None) == getattr(other, name, None) invalid in
a number of cases.  Unfortunately, there's no single meaning; some
people may want "loaded with an empty dict" to be different from
"not loaded" or "loaded as None" to be the same as "not loaded".

Better to refuse to guess than silently get it wrong for users.