Possibly a more drastic rewrite of docs #13
Comparing the previous documentation to other projects (especially
pytest, flask) I think I was trying to cover absolutely every
aspect, in as much detail as possible.  That seemed reasonable at
the time, because I'm familiar with the code.

Stepping back, it seems unapproachable in its sudden depth.  I'm
going to try another approach that walks through different options
as users might discover them.  That also means staying away from
"this is how DynamoDB works" as much as possible.  That can live in
a section in the future if it's something people want.