Fully test SearchIterator.__next__ #60
The last unit test, while large, is comprehensive over the possible
states that SearchIterator can be in when __next__ is called.

The table explains how the set of (page size, number of pages)
chains were selected.  To add a particular series of pages as a
test, just add a list of integers to the `chains` list at the end
of the function.