Shard, Coordinator no longer NamedTuple.  Buffer implements total ordering #17
Somehow I forgot that NamedTuple creates immutable classes, which is absolutely
not what I need.  For now they're really really dumb classes, but I may move
some logic back in.  TBD.

I also forgot that ApproximateCreationDateTime only has resolution to whole
seconds, so there will be many cases where heapq tries to order based on the
second attribute in the tuple, which is a dict.  That doesn't work.  Now, the
ordering element is (createtime, sequencenumber, clock).  The sequence number
isn't enough to guarantee ordering because there could be duplicates across
Shards, but the clock is an integer from the RecordBuffer that is guaranteed
to be unique and monotonically increasing.  In cases where two records from
two shards have the same create time and sequence number, we can't determine
the original ordering anyway, so we simply assign one.