Require models to be hashable, auto-discover parent hash functions
This one came up when I wanted to test equality of a dict by creating
identical instances of a model.  Because __eq__ was defined but
__hash__ wasn't, the WeakKeyDictionary bailed on the unhashable
subclass of BaseModel.

This is a big problem; the system requires Models to be hashable,
but there were no tests that tried to define __eq__ (thereby
giving up the default object.__hash__ function).

This makes that requirement explicit, and fails immediately if
a user tries to force a class to be unhashable (with __hash__ = None)