List, Set handle singular values. #73
This is fine for Set, which can only have primitive inner types.
However, List can have an inner type which is *also* list, and
then the gum and shoestring type check fails, causing heartache
and sadness.

This exists in the 1.0.0rc ec45ba442d2921729db18b6f1c0e3679c43b4663
and will be properly fixed before 1.0.0 releases.  If there is a 0.9.14,
it will not fix improve this fix.