Columns in base classes with the meta flag 'abstract' are now included in the derived class.
Fixed flake8 issues

Reversed list of bases to get oldest first
Added code from bug #72 as a test case

Updated to setup the bases after initializaion, and utulizing existing code.

Replaced type engine to register types by name instead of instance

The unit tests were getting messed up because of multiple instances of
the same model in the registry, and if you changed the order of anything,
it would get the wrong one.  This way, it stores the type by name, and
it went from 197 items to around 12 (something like that).