Remove engine.type_engine, #92
Type registration and binding are almost always going to use the built-in
bloop.Type._load and bloop.Type._dump, which removes 95% of the reason
for using declare's type engine.

This commit doesn't entirely remove the dependency on declare, although
it would still be a breaking change on its own (removing a public class's
public attribute).  If we want to merge this in, it should also remove
the remaining dependencies on declare:

  * declare.ModelMetaclass
  * declare.Field
  * declare.TypeDefinition

Finally if this will be a breaking change, we should adopt the
new python 3.6+ features __init_subclass__ and __set_name__ to
simplify the declarative model setup.  This will have the very
nice side effect of removing a metaclass, so users can include bloop
as a mixin for existing classes with their own metaclass.