Remove declare dependency, #92
This is the big one.  BaseModel starts using __init_subclass__,
and ModelMetaclass is removed entirely.  At the same time it was
simpler to remove Index._bind, which was a weird quasi-public
method.  The model_name property on Column, Index was simplified
but caused a bunch of churn in test code.

Still TODO:

  * So many docs changes, changelog and new migration section
  * See if this simplifies any of the work for column/index re-use
    in #92, or if it's a lot of breaking changes for nothing
  * Possibly simplify Type.dynamo_dump and Type.dynamo_load
    now that types aren't tightly coupled to a type engine?
    Still may be useful for cascading loads/dumps.