Automate Proxy subclass registration
bind_column, bind_index take optional proxy=False; when True, creates
an appropriately typed Proxy subclass around the given column/index.
If the column/index is already a proxy, it is first unwrapped to
minimize levels of indirection for getattr/setattr.

Because bind_column and bind_index may proxy the input object, these
methods now return the actual object that was bound to the model.