Add new Timestamp type and corresponding extensions, #87
Initially this will only have second-level resolution, although
there is the possibility to expose milli-/micro- as an init option.

If that's supported, the TTL integration should ensure that when
a bloop Timestamp is encountered, the resolution is seconds.  This
would be a similar special case to ensuring the bloop DateTime type
is not used.