Added `default` column argument
Applies a default value (or the return value of a function) to the column
on object creation.  The value is not persisted until saved.

Also added a validation Meta option.  Let me know if you want me to
seperate it out into another PR, but I felt that some validation was
in order, especially with default values.

pep8 fixes