fix sneaky test bug
The transaction test was failing to load a user's age for the model
  test/helpers/models.py::User

only when run with `tox -e unit` but not when run individually with:
  py.test tests/unit/test_transactions.py

Cutting down the test suites needed to repro, eventually this triggered it:

  py.test tests/unit/test_{models,transactions}.py

Attaching a debugger to the failing transaction test with:

  # previous line
  assert rx.items[0].obj.age == 3

  # with pdb
  try:
      assert rx.items[0].obj.age == 3
  except AssertionError:
      import pdb; pdb.set_trace()

I found the following when checking `User.Meta.columns`:

  > /home/crossj/ws/bloop/tests/unit/test_transactions.py(260)test_read_commit()
  -> session.transaction_read.assert_called_once_with(rx._request)
  (Pdb) User.Meta.columns
  {<Column[User.name]>, <Column[User.backups]>, <Column[User.joined]>, <Column[User.id=hash]>, <Column[User.email]>}

Note that there's a <Column[User.backups]> which doesn't exist.  When I checked
`User.Meta.columns_by_name` however, "age" was still a key.  So it wasn't the
bind methods in models.py doing something strange.

I went through and selectively disabled chunks of tests in test_models.py with '''
until two tests (there were three, but these were obvious) tripped it:

  def test_invalid_encryption(invalid_encryption):
  def test_invalid_backups(invalid_backups):

Both of these check that malformed values for Meta.{encryption,backups}
will fail as expected, and use a parameterized test value:

  @pytest.mark.parametrize("invalid_{}", [False, True, {}, User.age])

Note that the backups test comes second, and it seems that this structure causes
the column name to rebind:

  class Something:
      new_name = User.age

It turns out this is because a class's `__set_name__` can be called more than
once, whenever an outer class defines an attribute with this as its value.