Added token and move_to() for #132
The token property allows the caller to retrieve a Base64 encoded token that can be used with the move_to() method to resume a search. This works like the common AWS NextToken mechanism. Since Bloop's iterators work independently of DynamoDB request paging, the token includes both the DynamoDB LastEvaluatedKey and a page offset to enable Bloop to pick up exactly where it left off.